LUCENE-8837 smokeTestRelease.py option --download-only (#702)
* LUCENE-8837: smokeTestRelease.py option --download-only
Move download() and check_and() functions to scriptutil
Add cwd param to run() function in scriptutil
Move the check_ant function from buildAndPushRelease into scriptutil.py, and let it return the version.

(cherry picked from commit 44287d420624fe669a319f95f4712486bd57ae38)
